linting

+++ capacity/plot.py
@@ -1,9 +1,9 @@
 """ Plot -- make some handy plots """
 import csv
 
-import numpy as np
-
 from collections import defaultdict
+
+import numpy as np
 
 from matplotlib import pyplot as plt
 
@@ -17,7 +17,7 @@ def plot_cdf(data_list):
         sorted_x = np.sort(data)
         yvals = np.arange(len(sorted_x))/float(len(sorted_x))
         plt.plot(sorted_x, yvals, label=label)
-    
+
     plt.legend(frameon=False)
 
     plt.show()
@@ -31,13 +31,13 @@ def plot_traveler_time():
     with open(TRAVELER_STATS_FILE) as stat_file:
         reader = csv.reader(stat_file)
         for start, dest, wait, ride, transfer in reader:
-            mode_dict["wait"].append(int(wait))
-            mode_dict["ride"].append(int(ride))
+            mode_dict["wait"].append(float(wait))
+            mode_dict["ride"].append(float(ride))
             mode_dict["transfer"].append((transfer))
 
     plot_list = [
-                 ("Wait", mode_dict["wait"]),
-                 ("Ride", mode_dict["ride"]),
-                ]
-                 
+        ("Wait", mode_dict["wait"]),
+        ("Ride", mode_dict["ride"]),
+        ]
+
     plot_cdf(plot_list)
